+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 3.77.0 (November 8, 2022)
+  - show genres in autofilter-select
+  - allow deletion of filters
+
 ### 3.76.0 (September 1, 2022)
   - delete in online-reservation table if reservation in lupo exists for toy